requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fonttools==4.44.3
idna==3.4
kiwisolver==1.4.5
matplotlib==3.8.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
urllib3==2.1.0
